Refactor: 전역 스토어 기반 리팩터링
